[MOVE] account: Add dedicated files for account.fiscal.position model/views
